Fix `Dockerfile` for dev container (#1190)
The dev container is broken. I think this is from a later version of Rust, as it seems to work once we change from Ubuntu to Rust 1.75. I didn't test that it works on this repo, but I tested that it works [here](https://github.com/metaDAOproject/futarchy/blob/develop/.devcontainer/Dockerfile).